requirements.txt
azure-storage-blob
black
colormath
matplotlib
numpy
pillow==8.4.0
python-dotenv
requests
scikit-image
scikit-learn
scipy
selenium
tensorboard
torch
torch-tb-profiler
torchmetrics
torchvision